suspension model
file added is a potential model to use. We need to input the body mass, dampers and springs, not sure how to find these. SHould we calculate the values from the fusion doc? If yes i'm not sure between which points. 
the explanation of the model is available at: https://ctms.engin.umich.edu/CTMS/index.php?example=Suspension&section=SimulinkModeling